% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_variants.R
\name{load_variants}
\alias{load_variants}
\title{Load a variant call file.
Allowed types are either call_stats (MuTect1 format) or MuTect2 (VCF 3 and 4) format}
\usage{
load_variants(vcf, file_type = "vcf", reference, sample_name)
}
\arguments{
\item{file_type}{vcf or call_stats default is vcf}

\item{reference}{/path/to/reference Path to the reference genome used to call variants. Must be a BSGenome object}

\item{sample_name}{The sample name of the tumor in the vcf file}

\item{variant_file}{/path/to/variant/file}
}
\value{
Data frame with 5 columns: chr, start, end, ref_allele, alt_allele
}
\description{
The function is generally internal, but is exported to help the user in debugging input
}
\examples{
# Load variant calls from MuTect2
\dontrun{


}
}
