% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batcaver.R
\name{run_batcave}
\alias{run_batcave}
\title{Run the BATCAVE algorithm}
\usage{
run_batcave(vcf, reference, file_type = "vcf", seq_type = "wgs",
  sample_name = "TUMOR", min_vaf = 0.05, min_odds = 10,
  plot_path = NULL, profile_path = NULL)
}
\arguments{
\item{vcf}{Output of MuTect 1.1.7 (call stats) or 2.0 (VCF 4.0).}

\item{reference}{The genome reference used for alignment and variant calling (BSgenome Object)}

\item{file_type}{The file type of "vcf" (default = "vcf")}

\item{seq_type}{The type of sequencing experiment. Whole Genome "wgs" or Whole Exome "wes" (default = "wgs")}

\item{sample_name}{The name of the tumor sample (default = "TUMOR")}

\item{min_vaf}{The minimum allele frequency to use to compute mutation rate per base (default = "0.05)}

\item{min_odds}{The odds ratio cutoff for high-confidence mutations used to compute prior probability of mutation (default = "10" which corresponds to MuTect TLOD = 7.3)}

\item{plot_path}{The file name to use for a plot of the empirical mutation profile (default = "NULL" for no plot)}

\item{profile_path}{The file name to use for a tab separated file containing the empirical mutation profile (default = "NULL" for no file)}
}
\value{
a data frame with 11 columns: \code{chrom} the chromosome the variant is on,
\code{start} and \code{end} the position of the variant, \code{ref} and \code{alt} the reference and alternate alleles,
\code{context} the tri-nucleotide context of the variant, \code{TLOD} the score reported by MuTect, \code{freq} the variant allele frequency,
\code{pass_all} a boolean column that is TRUE is the variant passed all MuTect filters and FALSE otherwise,
\code{tlod_only} a boolean column that is TRUE is the variant passed all MuTect filters except the TLOD filter and false otherwise,
\code{pprob_variant} the posterior probability for the variant computed using the BATCAVE algorithm

@examples
\dontrun{
 ## input files and varianbles
 library(BSgenome.Hsapiens.UCSC.hg38) # the BSgenome reference
 vcf <- /path/to/MuTect/vcf
 reference <- BSgenome.Hsapiens.UCSC.hg38
 file_type <- "vcf"
 seq_type <- "wes"
 sample_name <- "TUMOR"
 min_vaf <- .1
 min_odds <- 10
 plot_path <- /path/to/empirical/profile/plot.pdf
 profile_path <- /path/to/empirical/profile/profile.tsv
 fr <- run_batcave(vcf = vcf, reference = reference, file_type = file_type,
                   seq_type = seq_type, sample_name = sample_name, min_vaf = min_vaf,
                   min_odds = min_odds, plot_path = plot_path, profile_path = profile_path)
 }
}
\description{
Run the BATCAVE algorithm
}
